third_party/blink/web_tests/fast/events/event-fire-disconnected-bubbling-mutation-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
Mutation events should propagate in disconnected DOM trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS eventReceivedOnParent is true
PASS eventReceivedOnChild is true
PASS successfullyParsed is true

TEST COMPLETE
